Split Client.getResponse to Client.doRequest & Client.getResponse.
There is some API that give empty response and we only need
to check HTTP status code.
example :
https://developer.github.com/v3/issues/assignees/#check-assignee

But there is no way to do this in current code because
getResponse will parse the response body and not returning
HTTP status code.

Client.doRequest makes it possible to do this.